changed type of radius from int to float

+++ Ball.cpp
@@ -1,7 +1,7 @@
 #include "Ball.h"
 
 
-Ball::Ball(int mNumber , int mRadius , float mMass , Vector2Df mPos)
+Ball::Ball(int mNumber , float mRadius , float mMass , Vector2Df mPos)
 {
 	this->number = mNumber;
 	this->mass = mMass;
